xgboost DMatrix generation

Geneartes a (list of) xgb.DMatrix. Supported for clusters. Requires Matrix and xgboost packages.

Laurae.xgb.dmat(data, label, missing = NA, save_names = NULL,
  save_keep = TRUE, clean_mem = FALSE, progress_bar = TRUE, ...)

Arguments

data
Type: matrix or dgCMatrix or data.frame or data.table or filename, or potentially a list of any of them. When a list is provided, it generates the appropriate xgb.DMatrix for all the sets. The data to convert to xgb.DMatrix. RAM usage required is 2x the current data input RAM usage, and 3x for data.frame and data.table due to internal matrix conversion added before binary matrix generation.

label
Type: numeric, or a list of numeric. The label of associated rows in data.

missing
Type: numeric. The value used to represent missing values in data. Defaults to NA (and missing values for dgCMatrix).

save_names
Type: character or NULL, or a list of characters. If names are provided, the generated xgb.DMatrix are stored physically to the drive. When a list is provided (along with a list of data and labels), it stores files sequentially by name if a list is provided for data but not for save_names. Defaults to NA.

save_keep
Type: logical, or a list of logicals. When names are provided, save_keep allows to selectively choose the xgb.DMatrix to retain for returning to the user. Useful when generating a list of xgb.DMatrix but choosing to keep only a part of them. When FALSE, it returns a NULL instead of the xgb.DMatrix. Defaults to TRUE.

clean_mem
Type: logical. Whether the force garbage collection at the end of each matrix construction in order to reclaim RAM. Defaults to FALSE.

progress_bar
Type: logical. Whether to print a progress bar in case of list inputs. Defaults to TRUE.

...
More arguments to pass to xgboost::xgb.DMatrix.

Value

The xgb.DMatrix

Examples

library(Matrix)
library(xgboost)

#> 
#> Attaching package: ‘xgboost’

#> The following objects are masked from ‘package:lightgbm’:
#> 
#>     getinfo, setinfo, slice


set.seed(0)

# Generate xgb.DMatrix from matrix
random_mat <- matrix(runif(1000000, 0, 1), nrow = 100000)
random_labels <- runif(100000, 0, 1)
xgb_from_mat <- Laurae.xgb.dmat(data = random_mat, label = random_labels, missing = NA)

# Generate xgb.DMatrix from data.frame
random_df <- data.frame(random_mat)
random_labels_2 <- runif(100000, 0, 1)
xgb_from_df <- Laurae.xgb.dmat(data = random_df, label = random_labels, missing = NA)

# Generate xgb.DMatrix from respective elements of a list with progress bar
# while keeping memory usage as low as theoretically possible
random_list <- list(random_mat, random_df)
random_labels_3 <- list(random_labels, random_labels_2)
xgb_from_list <- Laurae.xgb.dmat(data = random_list,
                                 label = random_labels_3,
                                 missing = NA,
                                 progress_bar = TRUE,
                                 clean_mem = TRUE)

#>    |                                                  | 0 % ~calculating     |+++++++++++++++++++++++++                         | 50% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s


# Generate xgb.DMatrix from respective elements of a list and keep only first
# while keeping memory usage as low as theoretically possible
xgb_from_list <- Laurae.xgb.dmat(data = random_list,
                                 label = random_labels_3,
                                 missing = NA,
                                 save_keep = c(TRUE, FALSE),
                                 clean_mem = TRUE)

#>    |                                                  | 0 % ~calculating     |+++++++++++++++++++++++++                         | 50% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s